heston_engine/requirements.txt
# Heston Stochastic Volatility Model Engine
# Requirements

# Core numerical libraries
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Web framework
flask>=2.0.0
flask-cors>=3.0.0

# Market data fetching
yfinance>=0.2.0
requests>=2.25.0

# Optional: FRED API for treasury rates
# fredapi>=0.5.0  # Requires free API key from https://fred.stlouisfed.org

# Optional: for matplotlib-based plots (not required for web UI)
matplotlib>=3.5.0

# Optional: benchmark validation vs QuantLib reference implementation
QuantLib-Python>=1.18
